requirements-docker.txt
fastapi==0.124.4
uvicorn==0.38.0
faiss-cpu==1.9.0
numpy==2.3.5
langchain-text-splitters==1.1.1
openai==2.13.0
python-dotenv==1.0.1
pydantic==2.12.5
